# openapi.yaml
openapi: 3.1.0
info:
  title: Unofficial Scalable Capital API
  version: 0.1.0
  description: |
    Local HTTP proxy that authenticates with Scalable Capital's private GraphQL API
    via Puppeteer interactive login and serves data to local clients.
servers:
  - url: http://127.0.0.1:3141
components:
  securitySchemes:
    GatewayToken:
      type: apiKey
      in: header
      name: X-Gateway-Token
      description: Optional gateway token set via --token CLI flag.
  schemas:
...
    SavingsTransaction:
      type: object
      required: [id, type, status, description, amount, currency, lastEventDateTime, cashTransactionType]
      properties:
        id:
          type: string
          example: "THEQVT1GGEOZJVNRTSMUII"
        type:
          type: string
          example: "CASH_TRANSACTION"
        status:
          type: string
          example: "SETTLED"
        description:
          type: string
          example: "Scalable Capital Tagesgeld Auszahlung"
        amount:
          type: number
          example: 2500
        currency:
          type: string
          example: "EUR"
        lastEventDateTime:
          type: string
          description: ISO-8601 datetime string.
          example: "2026-03-16T00:00:00.000Z"
        cashTransactionType:
          type: string
          enum: [WITHDRAWAL, CASH_TRANSFER_IN, CASH_TRANSFER_OUT, INTEREST_PAYOUT]
          example: "WITHDRAWAL"
    SavingsOverview:
      type: object
      properties:
        id:
          type: string
          description: The savings account ID.
          example: "2Yz9B847EvYmMdmJN68ywR"
        totalAmount:
          type: number
          description: Current account balance in EUR.
          example: 3400
        depositInterestRate:
          type: number
          description: Nominal annual interest rate as a decimal (e.g. 0.025 = 2.5%).
          example: 0.025
        nextPayoutDate:
          type: object
          properties:
            time:
              type: string
              description: ISO-8601 datetime of the next interest payout.
              example: "2026-04-01T00:00:00.000Z"
        interests:
          type: object
          properties:
            effectiveYearlyDepositInterestRate:
              type: number
              description: Effective annual yield as a decimal.
              example: 0.0252
            estimatedNextPayoutAmount:
              type: number
              description: Estimated interest amount to be paid out next.
              example: 3.9
            currentAccruedAmount:
              type: number
              description: Interest accrued so far in the current period.
              example: 1.34
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    GraphQLErrors:
      type: object
      description: Upstream GraphQL errors forwarded as-is.
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
...
    Transaction:
      type: object
      description: A single transaction (polymorphic — cash, security, non-trade security, or ELTIF).
      properties:
        id:
          type: string
        currency:
          type: string
        type:
          type: string
          enum: [CASH_TRANSACTION, SECURITY_TRANSACTION, NON_TRADE_SECURITY_TRANSACTION, ELTIF_TRANSACTION]
        status:
          type: string
          enum: [SETTLED, CANCELLED, PENDING]
        isCancellation:
          type: boolean
        lastEventDateTime:
          type: string
        description:
          type: string
        amount:
          type: number
          nullable: true
        isin:
          type: string
          nullable: true
        quantity:
          type: number
          nullable: true
        side:
          type: string
          enum: [BUY, SELL]
          nullable: true
        cashTransactionType:
          type: string
          nullable: true
        relatedIsin:
          type: string
          nullable: true
        securityTransactionType:
          type: string
          nullable: true
        nonTradeSecurityTransactionType:
          type: string
          nullable: true
    TransactionPage:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
          description: Pagination cursor for the next page; null when no more pages.
        total:
          type: number
          description: Total number of matching transactions.
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
paths:
...
  /transactions:
    get:
      summary: List transactions
      operationId: getTransactions
      description: |
        Returns a paginated list of transactions (trades, cash movements, transfers) from the broker portfolio.
        Filter by ISIN, type, status, or free-text search. Pass the returned `cursor` to fetch the next page.
      security:
        - GatewayToken: []
      parameters:
        - name: pageSize
          in: query
          schema:
            type: number
            default: 20
          description: Number of transactions to return.
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor from a previous response. Omit for the first page.
        - name: isin
          in: query
          schema:
            type: string
          description: Filter to transactions for a specific ISIN.
          example: NL0010273215
        - name: searchTerm
          in: query
          schema:
            type: string
          description: Free-text search across transaction descriptions.
        - name: type
          in: query
          schema:
            type: string
          description: Comma-separated list of transaction types to include.
          example: "SECURITY_TRANSACTION,CASH_TRANSACTION"
        - name: status
          in: query
          schema:
            type: string
          description: Comma-separated list of statuses to include.
          example: "SETTLED"
      responses:
        '200':
          description: Paginated transaction list
          content:
            application/json:
              schema:
                type: object
                properties:
                  account:
                    type: object
                    properties:
                      id:
                        type: string
                      brokerPortfolio:
                        type: object
                        properties:
                          id:
                            type: string
                          moreTransactions:
                            $ref: '#/components/schemas/TransactionPage'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Upstream GraphQL errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphQLErrors'
...
  /savings:
    get:
      summary: Savings account overview
      operationId: getSavings
      description: |
        Returns the current balance, nominal and effective interest rates, next payout date,
        and accrued interest for the Tagesgeld (overnight savings) account.
        Returns 503 if the session has no savings account.
      security:
        - GatewayToken: []
      responses:
        '200':
          description: Savings account overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsOverview'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Upstream GraphQL errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphQLErrors'
        '503':
          description: No savings account found in the current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /savings/transactions:
    get:
      summary: Savings account transactions
      operationId: getSavingsTransactions
      description: |
        Returns recent cash transactions for the Tagesgeld account (deposits, withdrawals,
        interest payouts). Uses the `OvernightOverviewPageData` GraphQL operation.
        Returns 503 if the session has no savings account.
      security:
        - GatewayToken: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of transactions to return.
      responses:
        '200':
          description: List of savings transactions
          content:
            application/json:
              schema:
                type: object
                required: [transactions]
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SavingsTransaction'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Upstream GraphQL errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphQLErrors'
        '503':
          description: No savings account found in the current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...